Unstable and untested release. Work in progress...

+++ Telemaco/src/com/diegomartin/telemaco/control/TripControl.java
@@ -27,7 +27,7 @@ public class TripControl {
 		TripDAO.update(trip);
 	}
 	
-	public static Trip readTrip(int id){
+	public static Trip readTrip(long id){
 		return TripDAO.read(id);
 	}
